requirements.txt
absl-py==2.0.0
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.7
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
docopt==0.6.2
executing==1.2.0
filelock==3.12.4
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
ipykernel==6.21.1
ipython==8.9.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpi4py==3.1.4
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.0
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.11
platformdirs==3.11.0
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.14.0
pyparsing==3.1.1
pyproject-toml==0.0.10
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
pyzmq==25.0.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
toml==0.10.2
torch==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.24.5
wcwidth==0.2.6
Werkzeug==3.0.0
wrapt==1.14.1
yarg==0.1.9
